+ now disable the use of precompiled shaders by default. Can be re-enabled with r__use_precompiled_shaders 1 in dev console = fix layered icons * enabled layered sounds in build_config_defines.h

+++ src/Layers/xrRender/ParticleEffectDef.h
@@ -39,8 +39,8 @@ namespace PS
 		}
 		IC void       		CalculateTC(int frame, Fvector2& lt, Fvector2& rb)
 		{
-			lt.x       	 	= (frame%m_iFrameDimX)*m_fTexSize.x;
-			lt.y        	= (frame/m_iFrameDimX)*m_fTexSize.y;
+			lt.x       	 	= (float)((frame%m_iFrameDimX)*m_fTexSize.x);
+			lt.y        	= (float)((frame/m_iFrameDimX)*m_fTexSize.y);
 			rb.x        	= lt.x+m_fTexSize.x;
 			rb.y        	= lt.y+m_fTexSize.y;
 		}
